created center styling component to reuse in routes file
