+++ App.tsx
@@ -1,25 +1,29 @@
-import React, { useState } from 'react';
-import { StyleSheet, Text, View, Button } from 'react-native';
-import SignUp from './components/SignUp';
-import SignIn from './components/SignIn';
+import { Routes } from './components/Routes'
+
+export default Routes
+
+// import React, { useState } from 'react';
+// import { StyleSheet, Text, View, Button } from 'react-native';
+// import SignUp from './components/SignUp';
+// import SignIn from './components/SignIn';
 
 
-export default function App() {
+// export default function App() {
 
-  return (
-    <View style={styles.container}>
-      {/* <Text>Hello World</Text> */}
-      <SignIn />
-      {/* <Button onPress={handleSignUp} /> */}
-    </View >
-  );
-}
+//   return (
+//     <View style={styles.container}>
+//       {/* <Text>Hello World</Text> */}
+//       <SignIn />
+//       {/* <Button onPress={handleSignUp} /> */}
+//     </View >
+//   );
+// }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+// const styles = StyleSheet.create({
+//   container: {
+//     flex: 1,
+//     backgroundColor: '#fff',
+//     alignItems: 'center',
+//     justifyContent: 'center',
+//   },
+// });
